S1P2213008-51 function for Find User password
