+++ backend/webmobile-project/src/main/java/com/ssafy/demo/controller/MemberController.java
@@ -111,49 +111,7 @@ public class MemberController {
 	}
 	
 	@GetMapping("findPassword/{email}")
-	public String mailSender(@PathVariable String email) throws AddressException, MessagingException, NoSuchAlgorithmException, UnsupportedEncodingException, GeneralSecurityException {
-		email = email.trim();
-		System.out.println("회원정보 :"+email);
-		MemberDTO member = service.selectByEmail(email);
-		if(member !=null && member.getEmail().equals(email)) {
-			// 네이버일 경우 smtp.naver.com 을 입력합니다.
-			// Google일 경우 smtp.gmail.com 을 입력합니다.
-			String host = "smtp.naver.com";
-			final String username = "qkrtlswhd"; // 네이버 아이디를 입력해주세요. @nave.com은 입력하지 마시구요.
-			final String password = "dkffkels32"; // 네이버 이메일 비밀번호를 입력해주세요.
-			int port = 465; // 포트번호
-			// 메일 내용
-			String recipient = member.getEmail(); // 받는 사람의 메일주소를 입력해주세요. starlight747@naver.com
-			String subject = member.getEmail()+" 회원 비밀번호"; // 메일 제목 입력해주세요. 
-			String body = member.getEmail()+"님의 비밀번호는 :"+(new AES256Util()).decrypt(member.getPassword()) +" 입니다.";
-			// 메일 내용 입력해주세요.
-			Properties props = System.getProperties(); // 정보를 담기 위한 객체 생성
-			// SMTP 서버 정보 설정
-			props.put("mail.smtp.host", host);
-			props.put("mail.smtp.port", port);
-			props.put("mail.smtp.auth", "true");
-			props.put("mail.smtp.ssl.enable", "true");
-			props.put("mail.smtp.ssl.trust", host); // Session 생성
-			Session session = Session.getDefaultInstance(props, new javax.mail.Authenticator() {
-				String un = username;
-				String pw = password;
-
-				protected javax.mail.PasswordAuthentication getPasswordAuthentication() {
-					return new javax.mail.PasswordAuthentication(un, pw);
-				}
-			});
-			session.setDebug(true); // for debug
-			Message mimeMessage = new MimeMessage(session); // MimeMessage 생성
-			mimeMessage.setFrom(new InternetAddress("qkrtlswhd@naver.com")); // 발신자 셋팅 , 보내는 사람의 이메일주소를 한번 더 입력합니다. 이때는 이메일 풀
-																			// 주소를 다 작성해주세요.
-			mimeMessage.setRecipient(Message.RecipientType.TO, new InternetAddress(recipient)); // 수신자셋팅 //.TO 외에 .CC(참조)
-																								// .BCC(숨은참조) 도 있음
-			mimeMessage.setSubject(subject); // 제목셋팅 
-			mimeMessage.setText(body); //내용셋팅
-			Transport.send(mimeMessage); // javax.mail.Transport.send() 이용
-			return "이메일 전송에 성공하였습니다.";
-		}else {
-			return "회원정보가 일치하지 않습니다";
-		}
+	public String mailSender(@PathVariable String email) throws AddressException, NoSuchAlgorithmException, UnsupportedEncodingException, MessagingException, GeneralSecurityException {
+		return service.mailSender(email);
 	}
 }
