+++ frontend/src/components/login/LoginPage.vue
@@ -65,7 +65,9 @@ export default {
         this.$store.state.user = user.data
         this.$session.start()
         this.$session.set('user', this.$store.state.user)
+        this.$session.set('seq', user.data.seq)
         alert(this.$session.get('user').name + '님 로그인에 성공하셨습니다.')
+        this.$router.go(0)
         this.cancel()
       } else {
         alert('회원정보가 일치하지 않습니다.')
