+++ frontend/src/views/PostDetailPage.vue
@@ -12,8 +12,7 @@
       <v-flex xs12 ml-8 mr-8 mb-4>
         <span class="title font-weight-thin">{{post.description}}</span>
       </v-flex>
-
-      <v-flex xs12 class="text-right"  v-if="this.$session.get('user')['seq'] === post.seq">
+      <v-flex xs12 class="text-right"  v-if="this.$session.get('seq') === post.seq">
         <v-btn class="font-weight-bold" color="teal accent-4" dark @click="updatePage">수정하기</v-btn>&nbsp;
         <v-btn  class="font-weight-bold" color="teal accent-4" dark @click="deletePost">삭제하기</v-btn>
       </v-flex>
@@ -57,6 +56,7 @@ export default {
         disqus_config: '',
         seq: ''
       },
+      seq: '',
       no: 0
     }
   },
@@ -64,6 +64,9 @@ export default {
 
   },
   mounted () {
+    // if(this.$session.get('user') != undefined){
+    //   this.seq = this.$session.get('user')['seq']
+    // }
     this.no = this.$route.params.no
     axios.get('http://192.168.31.114:8399/post/detailPost/' + this.no)
       .then(response => {
